Inlude font link 2

+++ Composer/ScriptHandler.php
@@ -21,12 +21,12 @@ class ScriptHandler
 
 			$ori_dir = "bundles/sopinetbootstrapextend/export/img";
 			$link = "web/img";
-			unlink($link);
+			if (file_exists($link)) unlink($link);
 			symlink($ori_dir, $link);
 
 			$ori_dir = "bundles/sopinetbootstrapextend/export/font";
 			$link = "web/font";
-			unlink($link);
+			if (file_exists($link)) unlink($link);
 			symlink($ori_dir, $link);
 
 		}
